### 2018-07-24 fb34834
Fixing bibtex error for JOSS
