+++ paper.bib
@@ -8,7 +8,7 @@
 	keywords = {OER, open educational resources, course evaluation, learning analytics, continuous improvement, The RISE Framework},
 	abstract = {The RISE (Resource Inspection, Selection, and Enhancement) Framework is a framework supporting the continuous improvement of open educational resources (OER).  The framework is an automated process that identifies learning resources that should be evaluated and either eliminated or improved.  This is particularly useful in OER contexts where the copyright permissions of resources allow for remixing, editing, and improving content.  The RISE Framework presents a scatterplot with resource usage on the x-axis and grade on the assessments associated with that resource on the y-axis.  This scatterplot is broken down into four different quadrants (the mean of each variable being the origin) to find resources that are candidates for improvement.  Resources that reside deep within their respective quadrant (farthest from the origin) should be further analyzed for continuous course improvement.  We present a case study applying our framework with an Introduction to Business course. Aggregate resource use data was collected from Google Analytics and aggregate assessment data was collected from an online assessment system.  Using the RISE Framework, we successfully identified resources, time periods, and modules in the course that should be further evaluated for improvement.},
   issn = {1492-3831},
-  url = {http://www.irrodl.org/index.php/irrodl/article/view/2952}
+  url = {http://www.irrodl.org/index.php/irrodl/article/view/2952}}
 
 @article{zenodo, 
   title={RISE package for R}, 
