### 2018-07-25 #846 @whedon
[REVIEW]: RISE: An R package for RISE analysis
**Submitting author:** @kalendar (<a href="http://orcid.org/0000-0001-6722-4744">David Wiley</a>)
**Repository:** <a href="https://github.com/lumenlearning/rise" target ="_blank">https://github.com/lumenlearning/rise</a>
**Version:** v1.0
