### 2018-07-30 b6c8b18
Changes to README
