### 2018-07-31 b238a10
Add build status and CRAN badges.
