### 2018-08-04 lumenlearning/rise#1 @seankross
[JOSS Review] Version Number
The CRAN version number and the version number in the DESCRIPTION file both match (which is good) however they do not match the latest GitHub release.

(https://github.com/openjournals/joss-reviews/issues/846)
