### 2018-08-05 lumenlearning/rise#1 @kalendar
Sean, what's the best way to resolve this problem?
