### 2018-08-06 lumenlearning/rise#1 @leeper
You can create a new release by running:

```bash
git tag -a [VERSION NUMBER] -m [VERSION NUMBER]
git push --tags
```

And editing the release/tag in the Github interface.
