### 2018-08-06 lumenlearning/rise#1 @kalendar
I'm sorry I wasn't clear in my previous comment. I know how to create a new release on GitHub. But even if I were to create a new release, the CRAN version and the latest GitHub release still wouldn't match? Is the goal to have the CRAN release and GitHub release be the same? In which case, I need to submit the updated version to CRAN? Or is there a different goal? Apologies for being a little slow on the uptake here.
