### 2018-08-06 lumenlearning/rise#1 @leeper
For JOSS, we want GitHub to be up to date so that when you make an archive of the repo at the end of the review process, that release matches what's gone through JOSS review. You can do this as the last stage of the review process.

A workflow I recommend is to make a GitHub release for every version that goes to CRAN so that it's easy to map CRAN releases back to GitHub commits. But ultimately what's on CRAN is not important from the JOSS perspective, though obviously you probably will update that once the JOSS review is complete.
