### 2018-08-06 lumenlearning/rise#2 @seankross
[JOSS Review] README Code Example
In the example code of your README, you use the function `rise_analysis()` where I think you mean to use the function `rise()`. See: https://github.com/lumenlearning/rise/blob/master/README.md#example

(openjournals/joss-reviews#846)
