### 2018-08-06 lumenlearning/rise#2 @kalendar
Good catch! Fixed.
