### 2018-08-06 lumenlearning/rise#3 @seankross
[JOSS Review] README DOI
[The citation](https://github.com/lumenlearning/rise#rise) at the beginning of the README should include the DOI for the paper it is referencing. (Providing the link along with the DOI is up to the author.)

(openjournals/joss-reviews#846)
