### 2018-08-07 lumenlearning/rise#3 @kalendar
Fixed.
