### 2018-08-07 #846 @seankross
@leeper my review of this package is complete and I recommend it for acceptance.
